third_party/blink/web_tests/http/tests/devtools/runtime/runtime-getProperties-expected.txt

Tests RemoteObject.getProperties.


Running: testSetUp

Running: testGetterAndSetter
{
    enumerable : true
    getter : {
        description : "get foo() { return 1; }"
        objectId : <string>
        type : "function"
    }
    isOwn : true
    name : "foo"
    private : false
    setter : {
        description : "set foo(value) { }"
        objectId : <string>
        type : "function"
    }
    synthetic : false
    wasThrown : false
    writable : false
}

Running: testGetterOnly
{
    enumerable : true
    getter : {
        description : "get foo() { return 1; }"
        objectId : <string>
        type : "function"
    }
    isOwn : true
    name : "foo"
    private : false
    synthetic : false
    wasThrown : false
    writable : false
}
